+++ pcs/cli/routing/cluster.py
@@ -84,8 +84,8 @@ cluster_cmd = create_router(
         "destroy": cluster.cluster_destroy,
         "verify": cluster.cluster_verify,
         "report": cluster.cluster_report,
-        "rename": cluster_command.cluster_rename,
         "remove_nodes_from_cib": cluster.remove_nodes_from_cib,
+        "rename": cluster_command.cluster_rename,
     },
     ["cluster"],
 )