+++ pcs_test/tier0/lib/pacemaker/test_live.py
@@ -1,5 +1,4 @@
 # pylint: disable=too-many-lines
-import os
 from unittest import (
     TestCase,
     mock,
@@ -327,10 +326,7 @@ class GetClusterStatusText(TestCase):
         )
 
 
-class GetCibXmlTestMixin:
-    def _call_command(self, runner, scope=None):
-        raise NotImplementedError()
-
+class GetCibXmlTest(TestCase):
     def test_success(self):
         expected_stdout = "<xml />"
         expected_stderr = ""
@@ -339,11 +335,10 @@ class GetCibXmlTestMixin:
             expected_stdout, expected_stderr, expected_retval
         )
 
-        real_xml = self._call_command(mock_runner)
+        real_xml = lib.get_cib_xml(mock_runner)
 
         mock_runner.run.assert_called_once_with(
-            [settings.cibadmin_exec, "--local", "--query"],
-            env_extend=self.env_extend,
+            [settings.cibadmin_exec, "--local", "--query"]
         )
         self.assertEqual(expected_stdout, real_xml)
 
@@ -357,7 +352,7 @@ class GetCibXmlTestMixin:
         )
 
         assert_raise_library_error(
-            lambda: self._call_command(mock_runner),
+            lambda: lib.get_cib_xml(mock_runner),
             (
                 Severity.ERROR,
                 report_codes.CIB_LOAD_ERROR,
@@ -368,8 +363,7 @@ class GetCibXmlTestMixin:
         )
 
         mock_runner.run.assert_called_once_with(
-            [settings.cibadmin_exec, "--local", "--query"],
-            env_extend=self.env_extend,
+            [settings.cibadmin_exec, "--local", "--query"]
         )
 
     def test_success_scope(self):
@@ -381,7 +375,7 @@ class GetCibXmlTestMixin:
             expected_stdout, expected_stderr, expected_retval
         )
 
-        real_xml = self._call_command(mock_runner, scope)
+        real_xml = lib.get_cib_xml(mock_runner, scope)
 
         mock_runner.run.assert_called_once_with(
             [
@@ -389,8 +383,7 @@ class GetCibXmlTestMixin:
                 "--local",
                 "--query",
                 "--scope={0}".format(scope),
-            ],
-            env_extend=self.env_extend,
+            ]
         )
         self.assertEqual(expected_stdout, real_xml)
 
@@ -409,7 +402,7 @@ class GetCibXmlTestMixin:
         )
 
         assert_raise_library_error(
-            lambda: self._call_command(mock_runner, scope=scope),
+            lambda: lib.get_cib_xml(mock_runner, scope=scope),
             (
                 Severity.ERROR,
                 report_codes.CIB_LOAD_ERROR_SCOPE_MISSING,
@@ -426,23 +419,8 @@ class GetCibXmlTestMixin:
                 "--local",
                 "--query",
                 "--scope={0}".format(scope),
-            ],
-            env_extend=self.env_extend,
+            ]
         )
-
-
-class GetCibXmlTest(GetCibXmlTestMixin, TestCase):
-    env_extend = None
-
-    def _call_command(self, runner, scope=None):
-        return lib.get_cib_xml(runner, scope)
-
-
-class GetCibDirectXmlTest(GetCibXmlTestMixin, TestCase):
-    env_extend = {"CIB_file": os.path.join(settings.cib_dir, "cib.xml")}
-
-    def _call_command(self, runner, scope=None):
-        return lib.get_cib_direct_xml(runner, scope)
 
 
 class GetCibTest(TestCase):